+++ NEWS.md
@@ -1,5 +1,7 @@
 # coro (development version)
 
+* Fixed performance issue related to JIT compilation of generator/async instances (#71). The JIT compiler is now able to properly cache the instances, making repeated instantiations fast, and preventing the JIT from kicking in on every call.
+
 * Async functions and generators can now be R6 methods (#63).
 
 * A memory leak that crept back in (see #36) was fixed with a more robust approach.